Rename index -> transactions
